chore: break up main.rs into smaller files
I chose to split along these lines:

- the machinery of turning flags into settings is now in src/settings.rs
- node rendering is now in src/render.rs

I didn't move the package.json parsing or git checks, as these
seem like major logic that belongs in the main file.

ALSO, renamed the cli from `ludwig` to `boltzmann`.
