+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 authors = ["C J Silverio <ceejceej@gmail.com>", "Chris Dickinson <chris@neversaw.us>"]
 edition = "2018"
-name = "ludwig"
+name = "boltzmann"
 version = "0.1.0"
 
 [dependencies]
